\name{setTISEANpath}
\alias{setTISEANpath}
\title{set TISEAN binaries folder path}
\description{set TISEAN binaries folder path.}
\usage{
setTISEANpath(path, GUI = interactive())
}
\arguments{
  \item{path}{path to TISEAN binaries folder}
  \item{GUI}{should a GUI be used?}
}
\value{the path itself}
\note{A text file storing the specified path will be stored in the
  user's HOME directory as a side effect.}
\keyword{ environment }
\keyword{ misc }
