\name{internals}
\alias{internals}
\alias{colnamesout}
\alias{as.character.TISEANoutput}
\alias{as.list.TISEANoutput}
\alias{as.matrixList}
\alias{as.matrix.TISEANblock}
\alias{as.matrix.TISEANoutput}
\alias{callNativeTISEAN}
\alias{callTISEAN}
\alias{helpTISEAN}
\alias{print.TISEANblock}
\alias{print.TISEANoutput}
\alias{setTISEANdocs}
\alias{TISEANblock}
\alias{TISEANoutput}
\title{Internal objects}
\description{Internal objects}
\keyword{internal}